pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "stockpulse"
version = "0.1.0"
description = "Automated Stock Scanning & Trading Signal System"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Martin", email = "martin@example.com"}
]

dependencies = [
    "yfinance>=0.2.36",
    "pandas>=2.1.0",
    "numpy>=1.24.0",
    "duckdb>=0.9.0",
    "apscheduler>=3.10.0",
    "pandas-ta>=0.3.14b",
    "plotly>=5.18.0",
    "streamlit>=1.29.0",
    "pyyaml>=6.0",
    "requests>=2.31.0",
    "pytz>=2023.3",
    "python-dateutil>=2.8.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
stockpulse = "stockpulse.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src/stockpulse"]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W"]
